[SofaDefaultType] move AbstractTypeInfo from DataTypeInfo.h to AbstractTypeInfo.h
